refactor(web): split landing sections and scope motion hooks
The landing page previously kept all section markup in one 3,000-line component and coordinated every animation through one stringly-typed controller. That made isolated review and lifecycle ownership difficult even though the rendered experience was stable.

Extract the unchanged section markup into purpose-specific server components. Replace the controller class with typed, section-scoped hooks, IntersectionObserver triggers, and explicit cleanup while preserving the existing SSR, reduced-motion, animation, and responsive behavior.
