+++ apps/web/app/_components/landing-motion.tsx
@@ -1,277 +1,521 @@
 "use client";
 
-import { useLayoutEffect } from "react";
+import { useLayoutEffect, useRef, type CSSProperties, type ReactNode, type RefObject } from "react";
 import { getRaceFrame, type RaceFrame, type RaceSeconds } from "../lib/landing-race";
 
-type Accent = { r: number; g: number; b: number };
+type RootRef = RefObject<HTMLDivElement | null>;
 type CopyStatus = "copied" | "error";
+type LandingElementName =
+  | "conn"
+  | "deployGrid"
+  | "dcell"
+  | "globe"
+  | "hero"
+  | "heroBg"
+  | "heroBottom"
+  | "heroTop"
+  | "l1"
+  | "l2"
+  | "mq"
+  | "nextFill"
+  | "nextLabel"
+  | "nextTime"
+  | "payoff"
+  | "plate"
+  | "plateGhost"
+  | "plateSub"
+  | "raceOuter"
+  | "ready"
+  | "scrollHint"
+  | "swapOuter"
+  | "vinextDone"
+  | "vinextFill"
+  | "vinextTime"
+  | "viteLabel";
 
-class Landing {
-  readonly props = { accent: "#f6821f", motion: true };
-  /** Median production-build times driving the benchmark race; the server
-      page passes measured values, these are only the pre-hydration default. */
-  race: RaceSeconds = { vinext: 3.1, nextjs: 6.2 };
-  root: HTMLElement | null = null;
-  reduce = false;
-  hero: HTMLElement | null = null;
-  heroBg: HTMLElement | null = null;
-  l1: HTMLElement | null = null;
-  l2: HTMLElement | null = null;
-  heroTop: HTMLElement | null = null;
-  heroBottom: HTMLElement | null = null;
-  scrollHint: HTMLElement | null = null;
-  ready: HTMLElement | null = null;
-  mq: HTMLElement | null = null;
-  swapOuter: HTMLElement | null = null;
-  nextLabel: HTMLElement | null = null;
-  viteLabel: HTMLElement | null = null;
-  plate: HTMLElement | null = null;
-  plateGhost: HTMLElement | null = null;
-  plateSub: HTMLElement | null = null;
-  conn: HTMLElement | null = null;
-  raceOuter: HTMLElement | null = null;
-  vFill: HTMLElement | null = null;
-  nFill: HTMLElement | null = null;
-  vTime: HTMLElement | null = null;
-  nTime: HTMLElement | null = null;
-  vDone: HTMLElement | null = null;
-  payoff: HTMLElement | null = null;
-  deployGrid: HTMLElement | null = null;
-  dcells: HTMLElement[] = [];
-  globe: HTMLElement | null = null;
-  gpins: HTMLElement[] = [];
-  _rv: HTMLElement[] = [];
-  _intro: HTMLElement[] = [];
-  _raf: number | null = null;
-  _dead = false;
-  _lastNow = 0;
-  _sy = 0;
-  _prevSy = 0;
-  _swapRun = false;
-  _raceRun = false;
-  _deployRun = false;
-  _loop: ((now: number) => void) | null = null;
-  _wake: (() => void) | null = null;
-  _visibilityChange: (() => void) | null = null;
-  _copyHandler: ((event: MouseEvent) => void) | null = null;
-  _copyResetTimer: ReturnType<typeof setTimeout> | null = null;
-  _drawParticles: ((now: number, heroProgress: number) => void) | null = null;
+const ACCENT = "#f6821f";
 
-  componentDidMount() {
-    const root = this.root;
+function findElement<T extends HTMLElement = HTMLElement>(
+  root: HTMLElement,
+  name: LandingElementName,
+): T | null {
+  return root.querySelector<T>(`[data-el="${name}"]`);
+}
+
+function findElements<T extends HTMLElement = HTMLElement>(
+  root: HTMLElement,
+  name: LandingElementName,
+): T[] {
+  return [...root.querySelectorAll<T>(`[data-el="${name}"]`)];
+}
+
+function prefersReducedMotion() {
+  return window.matchMedia("(prefers-reduced-motion: reduce)").matches;
+}
+
+function clamp(value: number) {
+  return value < 0 ? 0 : value > 1 ? 1 : value;
+}
+
+function smooth(value: number) {
+  const clamped = clamp(value);
+  return clamped * clamped * (3 - 2 * clamped);
+}
+
+function reveal(element: HTMLElement) {
+  element.style.opacity = "1";
+  element.style.transform = "none";
+}
+
+function applyRaceFrame(root: HTMLElement, frame: RaceFrame) {
+  const vinextFill = findElement(root, "vinextFill");
+  const nextFill = findElement(root, "nextFill");
+  const vinextTime = findElement(root, "vinextTime");
+  const nextTime = findElement(root, "nextTime");
+  const vinextDone = findElement(root, "vinextDone");
+
+  if (vinextFill) vinextFill.style.transform = `scaleX(${frame.vinextFill.toFixed(4)})`;
+  if (nextFill) nextFill.style.transform = `scaleX(${frame.nextjsFill.toFixed(4)})`;
+  if (vinextTime) vinextTime.textContent = `${frame.vinextTime.toFixed(1)}s`;
+  if (nextTime) nextTime.textContent = `${frame.nextjsTime.toFixed(1)}s`;
+  if (vinextDone) vinextDone.style.opacity = frame.vinextDone ? "1" : "0";
+}
+
+function useMotionFoundation(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
     if (!root) return;
 
-    this.reduce =
-      !this.props.motion || window.matchMedia("(prefers-reduced-motion: reduce)").matches;
-    this.applyAccent();
-
-    const $ = (selector: string) => root.querySelector<HTMLElement>(selector);
-    const $$ = (selector: string) => [...root.querySelectorAll<HTMLElement>(selector)];
-
-    this.hero = $('[data-el="hero"]');
-    this.heroBg = $('[data-el="heroBg"]');
-    this.l1 = $('[data-el="l1"]');
-    this.l2 = $('[data-el="l2"]');
-    this.heroTop = $('[data-el="heroTop"]');
-    this.heroBottom = $('[data-el="heroBottom"]');
-    this.scrollHint = $('[data-el="scrollHint"]');
-    this.ready = $('[data-el="ready"]');
-    this.mq = $('[data-el="mq"]');
-    this.swapOuter = $('[data-el="swapOuter"]');
-    this.nextLabel = $('[data-el="nextLabel"]');
-    this.viteLabel = $('[data-el="viteLabel"]');
-    this.plate = $('[data-el="plate"]');
-    this.plateGhost = $('[data-el="plateGhost"]');
-    this.plateSub = $('[data-el="plateSub"]');
-    this.conn = $('[data-el="conn"]');
-    this.raceOuter = $('[data-el="raceOuter"]');
-    this.vFill = $('[data-el="vinextFill"]');
-    this.nFill = $('[data-el="nextFill"]');
-    this.vTime = $('[data-el="vinextTime"]');
-    this.nTime = $('[data-el="nextTime"]');
-    this.vDone = $('[data-el="vinextDone"]');
-    this.payoff = $('[data-el="payoff"]');
-    this.deployGrid = $('[data-el="deployGrid"]');
-    this.dcells = $$('[data-el="dcell"]');
-    this.globe = $('[data-el="globe"]');
-    this.gpins = $$("[data-gpin]");
-    this._rv = $$("[data-rv]");
-    this._intro = $$("[data-intro]");
-
-    this.setupCopy(root);
-
-    if (this.reduce) {
-      this.staticFinish();
-      return;
-    }
-
-    document.documentElement.classList.add("motion-ready");
-    this.plate?.classList.remove("is-swapped");
-    if (this.plateGhost) this.plateGhost.textContent = "Turbopack";
-    if (this.nextLabel) this.nextLabel.style.opacity = "1";
-    if (this.viteLabel) {
-      this.viteLabel.style.opacity = "0";
-      this.viteLabel.style.transform = "translateY(44px)";
-    }
-    if (this.plateSub) {
-      this.plateSub.textContent = "next build · next.js 16";
-      this.plateSub.style.color = "var(--mute)";
-    }
-    if (this.conn) this.conn.style.opacity = ".3";
-    this.applyRaceFrame(getRaceFrame(this.race, 0));
-    if (this.payoff) {
-      this.payoff.style.opacity = "0";
-      this.payoff.style.transform = "translateY(16px)";
-    }
-
-    requestAnimationFrame(() => {
-      requestAnimationFrame(() => {
-        this._intro.forEach((element) => {
-          element.style.opacity = "1";
-          element.style.transform = "translateY(0)";
-        });
-      });
-    });
-
-    this.setupParticles();
-    this.runHeroSwap();
-    this._sy = window.scrollY;
-    this._prevSy = this._sy;
-    this._loop = this.loop.bind(this);
-    this._wake = () => {
-      if (this._dead || document.hidden || this._raf !== null || !this._loop) return;
-      this._lastNow = performance.now();
-      this._raf = requestAnimationFrame(this._loop);
-    };
-    this._visibilityChange = () => {
-      if (document.hidden && this._raf !== null) {
-        cancelAnimationFrame(this._raf);
-        this._raf = null;
-      } else {
-        this._wake?.();
-      }
-    };
-
-    document.addEventListener("visibilitychange", this._visibilityChange);
-    window.addEventListener("scroll", this._wake, { passive: true });
-    window.addEventListener("resize", this._wake, { passive: true });
-    this._wake();
-  }
-
-  destroy() {
-    this._dead = true;
-    document.documentElement.classList.remove("motion-ready");
-    if (this._raf !== null) cancelAnimationFrame(this._raf);
-    if (this._visibilityChange) {
-      document.removeEventListener("visibilitychange", this._visibilityChange);
-    }
-    if (this._wake) {
-      window.removeEventListener("scroll", this._wake);
-      window.removeEventListener("resize", this._wake);
-    }
-    if (this._copyHandler && this.root) {
-      this.root.removeEventListener("click", this._copyHandler);
-    }
-    if (this._copyResetTimer !== null) clearTimeout(this._copyResetTimer);
-  }
-
-  hexToRgb(hex: string): Accent {
-    let normalized = hex.replace("#", "");
-    if (normalized.length === 3) {
-      normalized = normalized
-        .split("")
-        .map((character) => character + character)
-        .join("");
-    }
-    const numeric = Number.parseInt(normalized || "f6821f", 16);
-    return { r: (numeric >> 16) & 255, g: (numeric >> 8) & 255, b: numeric & 255 };
-  }
-
-  applyAccent() {
-    if (!this.root) return;
-    const hex = this.props.accent;
-    const color = this.hexToRgb(hex);
-    this.root.style.setProperty("--orange", hex);
-    this.root.style.setProperty("--orange-rgb", `${color.r},${color.g},${color.b}`);
-    this.root.style.setProperty(
+    const normalized = ACCENT.slice(1);
+    const numeric = Number.parseInt(normalized, 16);
+    const color = { r: (numeric >> 16) & 255, g: (numeric >> 8) & 255, b: numeric & 255 };
+    root.style.setProperty("--orange", ACCENT);
+    root.style.setProperty("--orange-rgb", `${color.r},${color.g},${color.b}`);
+    root.style.setProperty(
       "--amber",
       `rgb(${Math.min(255, color.r + 8)},${Math.min(255, color.g + 43)},${Math.min(255, color.b + 34)})`,
     );
-  }
 
-  clamp(value: number) {
-    return value < 0 ? 0 : value > 1 ? 1 : value;
-  }
+    if (prefersReducedMotion()) return;
+    document.documentElement.classList.add("motion-ready");
+    return () => document.documentElement.classList.remove("motion-ready");
+  }, [rootRef]);
+}
 
-  smooth(value: number) {
-    const clamped = this.clamp(value);
-    return clamped * clamped * (3 - 2 * clamped);
-  }
+function useIntroAndRevealMotion(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
 
-  staticFinish() {
-    this._intro.forEach((element) => {
-      element.style.opacity = "1";
-      element.style.transform = "none";
+    const intro = [...root.querySelectorAll<HTMLElement>("[data-intro]")];
+    const reveals = [...root.querySelectorAll<HTMLElement>("[data-rv]")];
+    const ready = findElement(root, "ready");
+    if (prefersReducedMotion()) {
+      intro.forEach(reveal);
+      reveals.forEach(reveal);
+      if (ready) ready.style.opacity = "1";
+      return;
+    }
+
+    let innerFrame = 0;
+    const outerFrame = requestAnimationFrame(() => {
+      innerFrame = requestAnimationFrame(() => intro.forEach(reveal));
     });
-    this._rv.forEach((element) => {
-      element.style.opacity = "1";
-      element.style.transform = "none";
-    });
-    if (this.viteLabel) {
-      this.viteLabel.style.opacity = "1";
-      this.viteLabel.style.transform = "none";
-    }
-    if (this.ready) this.ready.style.opacity = "1";
-    if (this.nextLabel) this.nextLabel.style.opacity = "0";
-    this.plate?.classList.add("is-swapped");
-    if (this.plateSub) {
-      this.plateSub.textContent = "vite + @vitejs/plugin-rsc";
-      this.plateSub.style.color = "var(--orange-soft)";
-    }
-    if (this.plateGhost) this.plateGhost.textContent = "Vite";
-    this.applyRaceFrame(getRaceFrame(this.race, 1));
-    if (this.payoff) {
-      this.payoff.style.opacity = "1";
-      this.payoff.style.transform = "none";
-    }
-    this.lightDeploy(true);
-    this.globe?.classList.add("tether-done");
-    if (this.heroBg) this.heroBg.style.opacity = "0";
-  }
-
-  runHeroSwap() {
-    setTimeout(() => {
-      if (!this._dead && this.ready) this.ready.style.opacity = "1";
+    const readyTimer = window.setTimeout(() => {
+      if (ready) ready.style.opacity = "1";
     }, 1100);
-  }
+    const observer = new IntersectionObserver(
+      (entries) => {
+        for (const entry of entries) {
+          if (!entry.isIntersecting || !(entry.target instanceof HTMLElement)) continue;
+          reveal(entry.target);
+          observer.unobserve(entry.target);
+        }
+      },
+      { rootMargin: "0px 0px -8% 0px" },
+    );
+    for (const element of reveals) {
+      if (element.getBoundingClientRect().top < window.innerHeight * 0.92) reveal(element);
+      else observer.observe(element);
+    }
 
-  applyRaceFrame(frame: RaceFrame) {
-    if (this.vFill) this.vFill.style.transform = `scaleX(${frame.vinextFill.toFixed(4)})`;
-    if (this.nFill) this.nFill.style.transform = `scaleX(${frame.nextjsFill.toFixed(4)})`;
-    if (this.vTime) this.vTime.textContent = `${frame.vinextTime.toFixed(1)}s`;
-    if (this.nTime) this.nTime.textContent = `${frame.nextjsTime.toFixed(1)}s`;
-    if (this.vDone) this.vDone.style.opacity = frame.vinextDone ? "1" : "0";
-  }
+    return () => {
+      cancelAnimationFrame(outerFrame);
+      cancelAnimationFrame(innerFrame);
+      clearTimeout(readyTimer);
+      observer.disconnect();
+    };
+  }, [rootRef]);
+}
 
-  setupCopy(root: HTMLElement) {
+function useHeroMotion(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
+
+    const heroBg = findElement(root, "heroBg");
+    if (prefersReducedMotion()) {
+      if (heroBg) heroBg.style.opacity = "0";
+      return;
+    }
+
+    const hero = findElement(root, "hero");
+    const lineOne = findElement(root, "l1");
+    const lineTwo = findElement(root, "l2");
+    const heroTop = findElement(root, "heroTop");
+    const heroBottom = findElement(root, "heroBottom");
+    const scrollHint = findElement(root, "scrollHint");
+    const marquee = findElement(root, "mq");
+    const globe = findElement(root, "globe");
+    const deployGrid = findElement(root, "deployGrid");
+    if (heroBg) heroBg.style.opacity = "1";
+
+    let frame: number | null = null;
+    let lastNow = 0;
+    let scrollY = window.scrollY;
+    let previousScrollY = scrollY;
+    let dead = false;
+
+    const loop = (now: number) => {
+      if (dead || document.hidden) {
+        frame = null;
+        return;
+      }
+
+      let delta = now - (lastNow || now);
+      lastNow = now;
+      if (delta <= 0 || delta > 100) delta = 16.7;
+      const target = window.scrollY;
+      scrollY += (target - scrollY) * (1 - Math.pow(0.86, delta / 16.7));
+      if (Math.abs(target - scrollY) < 0.05) scrollY = target;
+      const velocity = (scrollY - previousScrollY) * (16.7 / delta);
+      previousScrollY = scrollY;
+      const viewportHeight = window.innerHeight;
+
+      let heroProgress = 0;
+      if (hero) {
+        heroProgress = clamp(scrollY / Math.max(1, hero.offsetHeight - viewportHeight));
+        const eased = smooth(heroProgress);
+        const distance = eased * window.innerWidth * 0.55;
+        const opacity = String(Math.max(0, 1 - eased * eased * 1.15));
+        const scale = `scale(${(1 - eased * 0.08).toFixed(4)})`;
+        if (lineOne) {
+          lineOne.style.transform = `translate3d(${-distance}px,0,0) ${scale}`;
+          lineOne.style.opacity = opacity;
+        }
+        if (lineTwo) {
+          lineTwo.style.transform = `translate3d(${distance}px,0,0) ${scale}`;
+          lineTwo.style.opacity = opacity;
+        }
+        const fade = clamp(heroProgress * 1.6);
+        const fadeOpacity = String(1 - fade);
+        if (heroTop) {
+          heroTop.style.opacity = fadeOpacity;
+          heroTop.style.transform = `translate3d(0,${-fade * 70}px,0)`;
+        }
+        if (heroBottom) {
+          heroBottom.style.opacity = fadeOpacity;
+          heroBottom.style.transform = `translate3d(0,${fade * 70}px,0)`;
+        }
+        if (scrollHint) scrollHint.style.opacity = String(1 - clamp(heroProgress * 4));
+      }
+
+      if (marquee) {
+        const relativeTop = marquee.getBoundingClientRect().top - viewportHeight;
+        const skew = Math.max(-5, Math.min(5, velocity * 0.08));
+        marquee.style.transform = `translate3d(${relativeTop * 0.42}px,0,0) skewX(${skew}deg)`;
+      }
+      if (globe && deployGrid) {
+        const rect = deployGrid.getBoundingClientRect();
+        const progress = clamp((viewportHeight - rect.top) / (viewportHeight + rect.height));
+        globe.style.transform = `translate3d(0,${((0.5 - progress) * 70).toFixed(1)}px,0)`;
+      }
+      if (heroBg) {
+        const progress = clamp(heroProgress * 1.15);
+        heroBg.style.opacity = String(1 - progress * progress * (3 - 2 * progress));
+      }
+
+      if (Math.abs(target - scrollY) > 0.05 || Math.abs(velocity) > 0.01) {
+        frame = requestAnimationFrame(loop);
+      } else {
+        frame = null;
+      }
+    };
+
+    const wake = () => {
+      if (dead || document.hidden || frame !== null) return;
+      lastNow = performance.now();
+      frame = requestAnimationFrame(loop);
+    };
+    const onVisibilityChange = () => {
+      if (document.hidden && frame !== null) {
+        cancelAnimationFrame(frame);
+        frame = null;
+      } else {
+        wake();
+      }
+    };
+
+    document.addEventListener("visibilitychange", onVisibilityChange);
+    window.addEventListener("scroll", wake, { passive: true });
+    window.addEventListener("resize", wake, { passive: true });
+    wake();
+    return () => {
+      dead = true;
+      if (frame !== null) cancelAnimationFrame(frame);
+      document.removeEventListener("visibilitychange", onVisibilityChange);
+      window.removeEventListener("scroll", wake);
+      window.removeEventListener("resize", wake);
+    };
+  }, [rootRef]);
+}
+
+function useEngineSwapMotion(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
+
+    const outer = findElement(root, "swapOuter");
+    const nextLabel = findElement(root, "nextLabel");
+    const viteLabel = findElement(root, "viteLabel");
+    const plate = findElement(root, "plate");
+    const plateGhost = findElement(root, "plateGhost");
+    const plateSub = findElement(root, "plateSub");
+    const connector = findElement(root, "conn");
+
+    const applySwap = (progress: number) => {
+      if (nextLabel) nextLabel.style.opacity = String(1 - Math.min(1, progress * 1.25));
+      if (viteLabel) {
+        viteLabel.style.opacity = String(Math.max(0, (progress - 0.15) / 0.85));
+        viteLabel.style.transform = `translateY(${(1 - progress) * 44}px)`;
+      }
+      if (plateSub) {
+        const swapped = progress > 0.5;
+        plateSub.textContent = swapped ? "vite + @vitejs/plugin-rsc" : "next build · next.js 16";
+        plateSub.style.color = swapped ? "var(--orange-soft)" : "var(--mute)";
+      }
+      if (connector) connector.style.opacity = String(0.3 + progress * 0.7);
+    };
+    const finish = () => {
+      if (viteLabel) reveal(viteLabel);
+      if (nextLabel) nextLabel.style.opacity = "0";
+      plate?.classList.add("is-swapped");
+      if (plateSub) {
+        plateSub.textContent = "vite + @vitejs/plugin-rsc";
+        plateSub.style.color = "var(--orange-soft)";
+      }
+      if (plateGhost) plateGhost.textContent = "Vite";
+      if (connector) connector.style.opacity = "1";
+    };
+    if (prefersReducedMotion()) {
+      finish();
+      return;
+    }
+
+    plate?.classList.remove("is-swapped");
+    if (plateGhost) plateGhost.textContent = "Turbopack";
+    if (nextLabel) nextLabel.style.opacity = "1";
+    if (viteLabel) {
+      viteLabel.style.opacity = "0";
+      viteLabel.style.transform = "translateY(44px)";
+    }
+    if (plateSub) {
+      plateSub.textContent = "next build · next.js 16";
+      plateSub.style.color = "var(--mute)";
+    }
+    if (connector) connector.style.opacity = ".3";
+
+    let animationFrame: number | null = null;
+    let widthTimer: number | null = null;
+    let started = false;
+    let dead = false;
+    const contractPlate = () => {
+      if (!plate || !plateGhost) return;
+      const initialWidth = plate.offsetWidth;
+      plateGhost.textContent = "Vite";
+      const finalWidth = plate.offsetWidth;
+      if (finalWidth === initialWidth) return;
+      plate.style.width = `${initialWidth}px`;
+      void plate.offsetWidth;
+      plate.style.transition = "width .6s var(--ease-out)";
+      plate.style.width = `${finalWidth}px`;
+      widthTimer = window.setTimeout(() => {
+        plate.style.width = "";
+        plate.style.transition = "";
+      }, 700);
+    };
+    const run = () => {
+      if (started) return;
+      started = true;
+      plate?.classList.add("is-swapped");
+      const start = performance.now();
+      const step = (now: number) => {
+        if (dead) return;
+        const time = Math.min(1, (now - start) / 1100);
+        applySwap(smooth(time));
+        if (time < 1) animationFrame = requestAnimationFrame(step);
+        else contractPlate();
+      };
+      animationFrame = requestAnimationFrame(step);
+    };
+    const check = () => {
+      if (outer && outer.getBoundingClientRect().top < -window.innerHeight * 0.075) run();
+    };
+
+    window.addEventListener("scroll", check, { passive: true });
+    window.addEventListener("resize", check, { passive: true });
+    check();
+    return () => {
+      dead = true;
+      if (animationFrame !== null) cancelAnimationFrame(animationFrame);
+      if (widthTimer !== null) clearTimeout(widthTimer);
+      window.removeEventListener("scroll", check);
+      window.removeEventListener("resize", check);
+    };
+  }, [rootRef]);
+}
+
+function useRaceMotion(rootRef: RootRef, race: RaceSeconds) {
+  const vinextSeconds = race.vinext;
+  const nextjsSeconds = race.nextjs;
+
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
+
+    const raceOuter = findElement(root, "raceOuter");
+    const payoff = findElement(root, "payoff");
+    const values = { vinext: vinextSeconds, nextjs: nextjsSeconds };
+    if (prefersReducedMotion()) {
+      applyRaceFrame(root, getRaceFrame(values, 1));
+      if (payoff) reveal(payoff);
+      return;
+    }
+
+    applyRaceFrame(root, getRaceFrame(values, 0));
+    if (payoff) {
+      payoff.style.opacity = "0";
+      payoff.style.transform = "translateY(16px)";
+    }
+
+    let frame: number | null = null;
+    let started = false;
+    let dead = false;
+    const run = () => {
+      if (started) return;
+      started = true;
+      const start = performance.now();
+      const duration = getRaceFrame(values, 0).durationMs;
+      const step = (now: number) => {
+        if (dead) return;
+        const time = Math.min(1, (now - start) / duration);
+        applyRaceFrame(root, getRaceFrame(values, time));
+        if (time < 1) frame = requestAnimationFrame(step);
+        else if (payoff) reveal(payoff);
+      };
+      frame = requestAnimationFrame(step);
+    };
+    if (!raceOuter) return;
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries.some((entry) => entry.isIntersecting)) {
+          run();
+          observer.disconnect();
+        }
+      },
+      { rootMargin: "0px 0px -45% 0px" },
+    );
+    observer.observe(raceOuter);
+    if (raceOuter.getBoundingClientRect().top < window.innerHeight * 0.55) run();
+
+    return () => {
+      dead = true;
+      observer.disconnect();
+      if (frame !== null) cancelAnimationFrame(frame);
+    };
+  }, [nextjsSeconds, rootRef, vinextSeconds]);
+}
+
+function illuminateDeploy(root: HTMLElement, instant: boolean) {
+  const timers: number[] = [];
+  root.querySelectorAll<HTMLElement>("[data-gpin]").forEach(reveal);
+  findElements(root, "dcell").forEach((cell, index) => {
+    const illuminate = () => {
+      const name = cell.querySelector<HTMLElement>("[data-name]");
+      const tag = cell.querySelector<HTMLElement>("[data-tag]");
+      if (name) name.style.color = "var(--ink)";
+      if (tag) tag.style.color = index === 0 ? "var(--orange-soft)" : "var(--ink-sub)";
+    };
+    if (instant) illuminate();
+    else timers.push(window.setTimeout(illuminate, index * 110));
+  });
+  return () => timers.forEach(clearTimeout);
+}
+
+function useDeployMotion(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
+
+    const deployGrid = findElement(root, "deployGrid");
+    const globe = findElement(root, "globe");
+    if (prefersReducedMotion()) {
+      illuminateDeploy(root, true);
+      globe?.classList.add("tether-done");
+      return;
+    }
+    if (!deployGrid) return;
+
+    let stopTimers = () => {};
+    let started = false;
+    const run = () => {
+      if (started) return;
+      started = true;
+      stopTimers = illuminateDeploy(root, false);
+      globe?.classList.add("tether-run");
+    };
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries.some((entry) => entry.isIntersecting)) {
+          run();
+          observer.disconnect();
+        }
+      },
+      { rootMargin: "0px 0px -20% 0px" },
+    );
+    observer.observe(deployGrid);
+    if (deployGrid.getBoundingClientRect().top < window.innerHeight * 0.8) run();
+
+    return () => {
+      observer.disconnect();
+      stopTimers();
+    };
+  }, [rootRef]);
+}
+
+function useCopyCommand(rootRef: RootRef) {
+  useLayoutEffect(() => {
+    const root = rootRef.current;
+    if (!root) return;
+
+    let resetTimer: number | null = null;
     const setStatus = (
       button: HTMLButtonElement,
       message: string,
       state: CopyStatus,
       label: string,
     ) => {
-      if (this._copyResetTimer !== null) clearTimeout(this._copyResetTimer);
+      if (resetTimer !== null) clearTimeout(resetTimer);
       button.textContent = message;
       button.setAttribute("data-copy-state", state);
       button.setAttribute("aria-label", label);
-      this._copyResetTimer = setTimeout(() => {
+      resetTimer = window.setTimeout(() => {
         button.textContent = "copy";
         button.removeAttribute("data-copy-state");
         button.setAttribute("aria-label", "Copy command");
       }, 1500);
     };
-
-    this._copyHandler = (event) => {
+    const onClick = (event: MouseEvent) => {
       if (!(event.target instanceof Element)) return;
       const button = event.target.closest("[data-copy]");
       if (!(button instanceof HTMLButtonElement) || !root.contains(button)) return;
@@ -282,7 +526,6 @@ class Landing {
         setStatus(button, "press ⌘C", "error", "Copy command failed");
         return;
       }
-
       Promise.race([
         navigator.clipboard.writeText(text),
         new Promise<never>((_, reject) => {
@@ -294,236 +537,33 @@ class Landing {
       );
     };
 
-    root.addEventListener("click", this._copyHandler);
-  }
-
-  loop(now: number) {
-    if (this._dead || document.hidden) {
-      this._raf = null;
-      return;
-    }
-
-    let delta = now - (this._lastNow || now);
-    this._lastNow = now;
-    if (delta <= 0 || delta > 100) delta = 16.7;
-    const target = window.scrollY;
-    this._sy += (target - this._sy) * (1 - Math.pow(0.86, delta / 16.7));
-    if (Math.abs(target - this._sy) < 0.05) this._sy = target;
-    const velocity = (this._sy - this._prevSy) * (16.7 / delta);
-    this._prevSy = this._sy;
-    const viewportHeight = window.innerHeight;
-
-    let heroProgress = 0;
-    if (this.hero) {
-      const heroHeight = this.hero.offsetHeight - viewportHeight;
-      heroProgress = this.clamp(this._sy / Math.max(1, heroHeight));
-      const eased = this.smooth(heroProgress);
-      const distance = eased * window.innerWidth * 0.55;
-      const opacity = String(Math.max(0, 1 - eased * eased * 1.15));
-      const scale = `scale(${(1 - eased * 0.08).toFixed(4)})`;
-      if (this.l1) {
-        this.l1.style.transform = `translate3d(${-distance}px,0,0) ${scale}`;
-        this.l1.style.opacity = opacity;
-      }
-      if (this.l2) {
-        this.l2.style.transform = `translate3d(${distance}px,0,0) ${scale}`;
-        this.l2.style.opacity = opacity;
-      }
-      const fade = this.clamp(heroProgress * 1.6);
-      const fadeOpacity = String(1 - fade);
-      if (this.heroTop) {
-        this.heroTop.style.opacity = fadeOpacity;
-        this.heroTop.style.transform = `translate3d(0,${-fade * 70}px,0)`;
-      }
-      if (this.heroBottom) {
-        this.heroBottom.style.opacity = fadeOpacity;
-        this.heroBottom.style.transform = `translate3d(0,${fade * 70}px,0)`;
-      }
-      if (this.scrollHint) {
-        this.scrollHint.style.opacity = String(1 - this.clamp(heroProgress * 4));
-      }
-    }
-
-    if (this.mq) {
-      const rect = this.mq.getBoundingClientRect();
-      const relativeTop = rect.top - viewportHeight;
-      const skew = Math.max(-5, Math.min(5, velocity * 0.08));
-      this.mq.style.transform = `translate3d(${relativeTop * 0.42}px,0,0) skewX(${skew}deg)`;
-    }
-
-    if (
-      !this._swapRun &&
-      this.swapOuter &&
-      this.swapOuter.getBoundingClientRect().top < -viewportHeight * 0.075
-    ) {
-      this.runSwap();
-    }
-
-    if (this._rv.length) {
-      this._rv = this._rv.filter((element) => {
-        if (element.getBoundingClientRect().top < viewportHeight * 0.92) {
-          element.style.opacity = "1";
-          element.style.transform = "none";
-          return false;
-        }
-        return true;
-      });
-    }
-
-    if (this.globe && this.deployGrid) {
-      const gridRect = this.deployGrid.getBoundingClientRect();
-      const progress = this.clamp(
-        (viewportHeight - gridRect.top) / (viewportHeight + gridRect.height),
-      );
-      this.globe.style.transform = `translate3d(0,${((0.5 - progress) * 70).toFixed(1)}px,0)`;
-    }
-
-    if (
-      !this._raceRun &&
-      this.raceOuter &&
-      this.raceOuter.getBoundingClientRect().top < viewportHeight * 0.55
-    ) {
-      this.runRace();
-    }
-    if (
-      !this._deployRun &&
-      this.deployGrid &&
-      this.deployGrid.getBoundingClientRect().top < viewportHeight * 0.8
-    ) {
-      this._deployRun = true;
-      this.lightDeploy(false);
-      this.globe?.classList.add("tether-run");
-    }
-
-    this._drawParticles?.(now, heroProgress);
-
-    if (Math.abs(target - this._sy) > 0.05 || Math.abs(velocity) > 0.01) {
-      if (this._loop) this._raf = requestAnimationFrame(this._loop);
-    } else {
-      this._raf = null;
-    }
-  }
-
-  runRace() {
-    if (this._raceRun) return;
-    this._raceRun = true;
-    const start = performance.now();
-    // Play the race in real time so the counters are honest stopwatches
-    // (linear, no easing — a second on screen is a build second). Capped so
-    // a slow ingest or a heavier benchmark suite can't drag the animation
-    // past attention span; beyond the cap it compresses proportionally.
-    const duration = getRaceFrame(this.race, 0).durationMs;
-
-    const step = (now: number) => {
-      if (this._dead) return;
-      const time = Math.min(1, (now - start) / duration);
-      const frame = getRaceFrame(this.race, time);
-      this.applyRaceFrame(frame);
-      if (time < 1) {
-        requestAnimationFrame(step);
-      } else if (this.payoff) {
-        this.payoff.style.opacity = "1";
-        this.payoff.style.transform = "none";
-      }
+    root.addEventListener("click", onClick);
+    return () => {
+      root.removeEventListener("click", onClick);
+      if (resetTimer !== null) clearTimeout(resetTimer);
     };
-
-    requestAnimationFrame(step);
-  }
-
-  applySwap(progress: number) {
-    if (this.nextLabel) {
-      this.nextLabel.style.opacity = String(1 - Math.min(1, progress * 1.25));
-    }
-    if (this.viteLabel) {
-      this.viteLabel.style.opacity = String(Math.max(0, (progress - 0.15) / 0.85));
-      this.viteLabel.style.transform = `translateY(${(1 - progress) * 44}px)`;
-    }
-    if (this.plateSub) {
-      const swapped = progress > 0.5;
-      this.plateSub.textContent = swapped ? "vite + @vitejs/plugin-rsc" : "next build · next.js 16";
-      this.plateSub.style.color = swapped ? "var(--orange-soft)" : "var(--mute)";
-    }
-    if (this.conn) this.conn.style.opacity = String(0.3 + progress * 0.7);
-  }
-
-  runSwap() {
-    if (this._swapRun) return;
-    this._swapRun = true;
-    this.plate?.classList.add("is-swapped");
-    const start = performance.now();
-    const duration = 1100;
-
-    const step = (now: number) => {
-      if (this._dead) return;
-      const time = Math.min(1, (now - start) / duration);
-      this.applySwap(this.smooth(time));
-      if (time < 1) requestAnimationFrame(step);
-      else this.contractPlate();
-    };
-
-    requestAnimationFrame(step);
-  }
-
-  contractPlate() {
-    if (!this.plate || !this.plateGhost) return;
-    const initialWidth = this.plate.offsetWidth;
-    this.plateGhost.textContent = "Vite";
-    const finalWidth = this.plate.offsetWidth;
-    if (finalWidth === initialWidth) return;
-    this.plate.style.width = `${initialWidth}px`;
-    void this.plate.offsetWidth;
-    this.plate.style.transition = "width .6s var(--ease-out)";
-    this.plate.style.width = `${finalWidth}px`;
-    const plate = this.plate;
-    setTimeout(() => {
-      plate.style.width = "";
-      plate.style.transition = "";
-    }, 700);
-  }
-
-  lightDeploy(instant: boolean) {
-    this.gpins.forEach((pin) => {
-      pin.style.opacity = "1";
-      pin.style.transform = "none";
-    });
-    this.dcells.forEach((cell, index) => {
-      const illuminate = () => {
-        const name = cell.querySelector<HTMLElement>("[data-name]");
-        const tag = cell.querySelector<HTMLElement>("[data-tag]");
-        if (name) name.style.color = "var(--ink)";
-        if (tag) tag.style.color = index === 0 ? "var(--orange-soft)" : "var(--ink-sub)";
-      };
-      if (instant) illuminate();
-      else setTimeout(illuminate, index * 110);
-    });
-  }
-
-  setupParticles() {
-    if (!this.heroBg) return;
-    const heroBg = this.heroBg;
-    heroBg.style.opacity = "1";
-    this._drawParticles = (_now, heroProgress) => {
-      const progress = this.clamp(heroProgress * 1.15);
-      heroBg.style.opacity = String(1 - progress * progress * (3 - 2 * progress));
-    };
-  }
+  }, [rootRef]);
 }
 
-export function LandingMotion({ race }: { race?: RaceSeconds }) {
-  const vinextSeconds = race?.vinext;
-  const nextjsSeconds = race?.nextjs;
+type LandingMotionProps = {
+  children: ReactNode;
+  race: RaceSeconds;
+  style: CSSProperties;
+};
 
-  useLayoutEffect(() => {
-    const landing = new Landing();
-    if (vinextSeconds && nextjsSeconds) {
-      landing.race = { vinext: vinextSeconds, nextjs: nextjsSeconds };
-    }
-    landing.root = document.getElementById("app");
-    landing.componentDidMount();
-    return () => landing.destroy();
-    // Primitive deps: a fresh `race` object with equal values must not tear
-    // down and rebuild the scroll/animation controller.
-  }, [vinextSeconds, nextjsSeconds]);
+export function LandingMotion({ children, race, style }: LandingMotionProps) {
+  const rootRef = useRef<HTMLDivElement>(null);
+  useMotionFoundation(rootRef);
+  useIntroAndRevealMotion(rootRef);
+  useHeroMotion(rootRef);
+  useEngineSwapMotion(rootRef);
+  useRaceMotion(rootRef, race);
+  useDeployMotion(rootRef);
+  useCopyCommand(rootRef);
 
-  return null;
+  return (
+    <div ref={rootRef} id="app" className="landing-root" style={style}>
+      {children}
+    </div>
+  );
 }
